fix some annoying css issues
