+++ typescript/src/renderer/components/PageDelimiter.tsx
@@ -11,9 +11,8 @@ export const PageDelimiter: React.FC<PageDelimiterProps> = ({
 }) => {
   return (
     <div className={`jsondoc-page-delimiter ${className || ""}`.trim()}>
-      <div className="jsondoc-page-delimiter-line">
-        <span className="jsondoc-page-number">Page {pageNumber}</span>
-      </div>
+      <div className="jsondoc-page-delimiter-line"></div>
+      <span className="jsondoc-page-number">Page {pageNumber}</span>
     </div>
   );
 };
